Add MMC/EMMC bus width quirks.
This is very hacky solution for a problem with a bus width switching
between MMC and EMMC card. In general U-Boot MMC driver was not supposed
to reinitialize the MMC controller. The problem is that the read-only
card capabilities are set (also based on hardcoded bus width value) in
controller initialization function (what is somehow logical as in
general the maximum bus width on HW shouldn't change in the runtime).
So, as we know that for EMMC we always have 8 bit and for MMC we always
have 4 bit width - we can perform a little hack and change the card
capabilities right before the bus width is being set.