drivers: mmc: fix reinitialization error
This commit fixes the re-initialization error. If the initialization
failed once it will fail always with a timeout on MMC controller
transmission. For some reason calling the controller init function while
there was no card present was getting it to some unknown state. It's not
necessary to reinitialize the controller every time we rescan for a MMC
card - the controller is initialized once on a startup and that's okay.