drivers: dw mmc: modify timeouts
Update timeouts in the dwmci_send_cmd function. The change is based on
the one done in U-Boot 2017.7. Having such a long timeouts is not really
necessary and is causing problems, e.g. the mmc rescan command is
hanging up when the AA1 mux is set to QSPI - this commit fixes that
issue.